ci: Be consistent when specifying cloudsmith repos.
The strategy elsewhere, already in use and documented is to
use CLOUDSMITH_STABLE_REPO and CLOUDSMITH_UNSTABLE_REPO
for the cloudsmith repo names. This defaults to the repos used
in the main branch, with possibilies for devs to set them
in the builder environment without messing with the sources.

I see no reason to change this. This commit applies this
strategy also to the travis raspbian builds.